(refactor) change page (views) vue file names
