+++ src/views/VueFundamentalsShoppingList.vue
